Refactor Assembler into Facade with SentenceState and ExpressionState
- Moved `src/semantic/assembler.rs` to `src/semantic/assembler/mod.rs`
- Extracted `SentenceState` logic to `src/semantic/assembler/sentence.rs`
- Extracted `ExpressionState` logic to `src/semantic/assembler/expression.rs`
- Extracted data models to `src/semantic/assembler/model.rs`
- Updated `.jules/atlas.md` with architectural journal entry
